trademonkey-fusion-backend/requirements.txt
# File: trademonkey-fusion-backend/requirements.txt
# Existing requirements
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
coloredlogs>=15.0.0
ta>=0.10.0
asyncio-mqtt>=0.13.0
aiohttp>=3.8.0

# NEW API Server Requirements
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0.0
redis>=5.0.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Sentiment Analysis Requirements
tweepy>=4.14.0
praw>=7.7.0
textblob>=0.17.1
nltk>=3.8.1

# Optional for production
gunicorn>=21.2.0